Update secp256r1 tests
